Rename actionDispatcher to actionCreator and remove dispatching from it
