+++ packages/satcheljs/lib/index.ts
@@ -1,7 +1,7 @@
 export { default as ActionCreator } from './interfaces/ActionCreator';
 export { default as ActionMessage } from './interfaces/ActionMessage';
 export { default as Subscriber } from './interfaces/Subscriber';
-export { default as actionDispatcher } from './actionDispatcher';
+export { default as actionCreator } from './actionCreator';
 export { default as mutator } from './mutator';
 export { default as orchestrator } from './orchestrator';
