+++ packages/satcheljs/lib/simpleAction.ts
@@ -1,11 +1,11 @@
 import SimpleAction from './interfaces/SimpleAction';
-import actionDispatcher from './actionDispatcher';
+import actionCreator from './actionCreator';
 import mutator from './mutator';
 
 export default function simpleAction(actionType: string) {
     return function createSimpleAction<T extends SimpleAction>(target: T): T {
         // Create an action dispatcher
-        let simpleActionDispatcher = actionDispatcher(actionType)(
+        let simpleActionCreator = actionCreator(actionType)(
             function simpleActionCreator() {
                 return {
                     type: actionType,
@@ -14,12 +14,12 @@ export default function simpleAction(actionType: string) {
             });
 
         // Create a mutator that subscribes to that action creator
-        mutator(simpleActionDispatcher)(
+        mutator(simpleActionCreator)(
             function simpleActionMutator(actionMessage) {
                 target.apply(null, actionMessage.args);
             });
 
         // Return a function that dispatches that action
-        return simpleActionDispatcher as any as T;
+        return simpleActionCreator as any as T;
     }
 }
